+++ tests/Notifier/CliBasedNotifierTestTrait.php
@@ -77,16 +77,20 @@ trait CliBasedNotifierTestTrait
             [
                 (new Notification())
                     ->setBody('I\'m the notification body')
-                    ->setTitle('I\'m the notification title')
                     ->addOption('subtitle', 'I\'m the notification subtitle'),
                 $this->getExpectedCommandLineForNotificationWithASubtitle(),
             ],
             [
                 (new Notification())
                     ->setBody('I\'m the notification body')
-                    ->setTitle('I\'m the notification title')
                     ->addOption('sound', 'Frog'),
                 $this->getExpectedCommandLineForNotificationWithASound(),
+            ],
+            [
+                (new Notification())
+                    ->setBody('I\'m the notification body')
+                    ->addOption('url', 'https://google.com'),
+                $this->getExpectedCommandLineForNotificationWithAnUrl(),
             ],
             [
                 (new Notification())
@@ -100,6 +104,7 @@ trait CliBasedNotifierTestTrait
                     ->setTitle('I\'m the notification title')
                     ->addOption('subtitle', 'I\'m the notification subtitle')
                     ->addOption('sound', 'Frog')
+                    ->addOption('url', 'https://google.com')
                     ->setIcon($iconDir.'/image.gif'),
                 $this->getExpectedCommandLineForNotificationWithAllOptions(),
             ],
@@ -157,7 +162,7 @@ trait CliBasedNotifierTestTrait
      */
     protected function getExpectedCommandLineForNotificationWithASubtitle()
     {
-        return $this->getExpectedCommandLineForNotificationWithATitle();
+        return $this->getExpectedCommandLineForNotification();
     }
 
     /**
@@ -167,7 +172,17 @@ trait CliBasedNotifierTestTrait
      */
     protected function getExpectedCommandLineForNotificationWithASound()
     {
-        return $this->getExpectedCommandLineForNotificationWithATitle();
+        return $this->getExpectedCommandLineForNotification();
+    }
+
+    /**
+     * Sound is supported only on few notifier.
+     *
+     * @return string
+     */
+    protected function getExpectedCommandLineForNotificationWithAnUrl()
+    {
+        return $this->getExpectedCommandLineForNotification();
     }
 
     /**